select: extract into module/select/ following the _socket pattern
importing.rs shrinks by 163 lines.
