+++ pyre/pyre-interpreter/src/importing.rs
@@ -141,7 +141,7 @@ pub fn install_builtin_modules() {
     register_builtin_module("fcntl", init_fcntl);
     #[cfg(unix)]
     register_builtin_module("syslog", init_syslog);
-    register_builtin_module("select", init_select);
+    register_builtin_module("select", crate::module::select::moduledef::init);
     register_builtin_module("termios", crate::module::termios::moduledef::init);
     register_builtin_module("_socket", crate::module::_socket::moduledef::init);
     register_builtin_module("mmap", crate::module::mmap::moduledef::init);
@@ -2508,170 +2508,6 @@ fn init_syslog(ns: &mut DictStorage) {
             1,
         ),
     );
-}
-
-/// _select module — PyPy: pypy/module/select/.
-///
-/// Implements `select.select(rlist, wlist, xlist, timeout=None)` via
-/// `rustpython_host_env::select::{FdSet, select, sec_to_timeval}`.  poll()
-/// / epoll / kqueue object types are not implemented yet; they need
-/// per-instance heap state which the current pyre builtin-module wiring
-/// doesn't expose.
-fn init_select(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "select",
-        crate::make_builtin_function("select", |args| {
-            #[cfg(all(unix, feature = "host_env"))]
-            {
-                use rustpython_host_env::select as host_select;
-
-                if args.len() < 3 {
-                    return Err(crate::PyError::type_error(
-                        "select() takes at least 3 arguments",
-                    ));
-                }
-
-                fn collect_fds(
-                    seq: pyre_object::PyObjectRef,
-                ) -> Result<Vec<(pyre_object::PyObjectRef, i32)>, crate::PyError> {
-                    unsafe {
-                        let is_list = pyre_object::is_list(seq);
-                        let is_tuple = pyre_object::is_tuple(seq);
-                        if !is_list && !is_tuple {
-                            return Err(crate::PyError::type_error(
-                                "select() arguments 1-3 must be sequences",
-                            ));
-                        }
-                        let n = if is_list {
-                            pyre_object::w_list_len(seq)
-                        } else {
-                            pyre_object::w_tuple_len(seq)
-                        };
-                        let mut out = Vec::with_capacity(n);
-                        for i in 0..n {
-                            let item = if is_list {
-                                pyre_object::w_list_getitem(seq, i as i64)
-                            } else {
-                                pyre_object::w_tuple_getitem(seq, i as i64)
-                            }
-                            .ok_or_else(|| {
-                                crate::PyError::value_error("select() sequence item missing")
-                            })?;
-                            if !pyre_object::is_int(item) {
-                                return Err(crate::PyError::type_error(
-                                    "argument must be an int, or have a fileno() method",
-                                ));
-                            }
-                            let fd = pyre_object::w_int_get_value(item) as i32;
-                            if fd < 0 {
-                                return Err(crate::PyError::value_error(
-                                    "file descriptor cannot be a negative integer",
-                                ));
-                            }
-                            out.push((item, fd));
-                        }
-                        Ok(out)
-                    }
-                }
-
-                let rfds = collect_fds(args[0])?;
-                let wfds = collect_fds(args[1])?;
-                let xfds = collect_fds(args[2])?;
-
-                let mut rset = host_select::FdSet::new();
-                let mut wset = host_select::FdSet::new();
-                let mut xset = host_select::FdSet::new();
-                let mut nfds: i32 = -1;
-                for &(_, fd) in &rfds {
-                    rset.insert(fd);
-                    if fd > nfds {
-                        nfds = fd;
-                    }
-                }
-                for &(_, fd) in &wfds {
-                    wset.insert(fd);
-                    if fd > nfds {
-                        nfds = fd;
-                    }
-                }
-                for &(_, fd) in &xfds {
-                    xset.insert(fd);
-                    if fd > nfds {
-                        nfds = fd;
-                    }
-                }
-
-                let mut tv_storage;
-                let timeout_ref: Option<&mut host_select::timeval> = match args.get(3) {
-                    None => None,
-                    Some(&t) if unsafe { pyre_object::is_none(t) } => None,
-                    Some(&t) => {
-                        let secs = unsafe {
-                            if pyre_object::is_float(t) {
-                                pyre_object::w_float_get_value(t)
-                            } else if pyre_object::is_int(t) {
-                                pyre_object::w_int_get_value(t) as f64
-                            } else {
-                                return Err(crate::PyError::type_error(
-                                    "timeout must be a float or None",
-                                ));
-                            }
-                        };
-                        if secs < 0.0 {
-                            return Err(crate::PyError::value_error(
-                                "timeout must be non-negative",
-                            ));
-                        }
-                        tv_storage = host_select::sec_to_timeval(secs);
-                        Some(&mut tv_storage)
-                    }
-                };
-
-                let n = host_select::select(nfds + 1, &mut rset, &mut wset, &mut xset, timeout_ref)
-                    .map_err(|e| {
-                        crate::PyError::os_error_with_errno(
-                            e.raw_os_error().unwrap_or(0),
-                            format!("select: {e}"),
-                        )
-                    })?;
-                let _ = n;
-
-                fn build_ready(
-                    set: &mut host_select::FdSet,
-                    inputs: &[(pyre_object::PyObjectRef, i32)],
-                ) -> pyre_object::PyObjectRef {
-                    let items: Vec<_> = inputs
-                        .iter()
-                        .filter_map(|&(obj, fd)| if set.contains(fd) { Some(obj) } else { None })
-                        .collect();
-                    pyre_object::w_list_new(items)
-                }
-
-                let r_ready = build_ready(&mut rset, &rfds);
-                let w_ready = build_ready(&mut wset, &wfds);
-                let x_ready = build_ready(&mut xset, &xfds);
-                Ok(pyre_object::w_tuple_new(vec![r_ready, w_ready, x_ready]))
-            }
-            #[cfg(not(all(unix, feature = "host_env")))]
-            {
-                let _ = args;
-                Err(crate::PyError::not_implemented(
-                    "select.select requires host_env feature on a Unix platform",
-                ))
-            }
-        }),
-    );
-
-    crate::dict_storage_store(ns, "error", pyre_object::w_str_new("OSError"));
-    #[cfg(unix)]
-    {
-        crate::dict_storage_store(
-            ns,
-            "PIPE_BUF",
-            pyre_object::w_int_new(libc::PIPE_BUF as i64),
-        );
-    }
 }
 
 /// itertools stub
