+++ pyre/pyre-interpreter/src/module/select/moduledef.rs
@@ -1,0 +1,7 @@
+//! select module definition — PyPy: pypy/module/select/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_select::register_module(ns);
+}
